+++ pyriemann/clustering.py
@@ -12,10 +12,10 @@ from .classification import MDM
 
 
 def _fit_single(X, y=None, n_clusters=2, init='random', random_state=None,
-                metric='riemann', max_iter=100, tol=1e-4):
+                metric='riemann', max_iter=100, tol=1e-4, n_jobs=1):
     """helper to fit a single run of centroid."""
     # init random state if provided
-    mdm = MDM(metric=metric)
+    mdm = MDM(metric=metric, n_jobs=n_jobs)
     squared_nomrs = [numpy.linalg.norm(x, ord='fro')**2 for x in X]
     mdm.covmeans_ = _init_centroids(X, n_clusters, init,
                                     random_state=random_state,
@@ -139,7 +139,8 @@ class Kmeans(BaseEstimator, ClassifierMixin, ClusterMixin, TransformerMixin):
                                                random_state=self.seed,
                                                metric=self.metric,
                                                max_iter=self.max_iter,
-                                               tol=self.tol)
+                                               tol=self.tol,
+                                               n_jobs=self.n_jobs)
         else:
             numpy.random.seed(self.seed)
             seeds = numpy.random.randint(
@@ -164,7 +165,8 @@ class Kmeans(BaseEstimator, ClassifierMixin, ClusterMixin, TransformerMixin):
                                          random_state=seed,
                                          metric=self.metric,
                                          max_iter=self.max_iter,
-                                         tol=self.tol)
+                                         tol=self.tol,
+                                         n_jobs=1)
                     for seed in seeds)
                 labels, inertia, mdm = zip(*res)
 
@@ -241,7 +243,7 @@ class KmeansPerClassTransform(BaseEstimator, TransformerMixin):
 
     def transform(self, X):
         """transform."""
-        mdm = MDM(metric=self.metric)
+        mdm = MDM(metric=self.metric, n_jobs=self.km.n_jobs)
         mdm.covmeans_ = self.covmeans_
         return mdm._predict_distances(X)
 
@@ -263,6 +265,10 @@ class Potato(BaseEstimator, TransformerMixin, ClassifierMixin):
         The number of standard deviation to reject artifacts.
     n_iter_max : int (default 100)
         The maximum number of iteration to reach convergence.
+    pos_label: int (default 1)
+        The positive label corresponding to clean data
+    neg_label: int (default 0)
+        The negative label corresponding to artifact data
 
     Notes
     -----
@@ -281,11 +287,16 @@ class Potato(BaseEstimator, TransformerMixin, ClassifierMixin):
     2013.
     """
 
-    def __init__(self, metric='riemann', threshold=3, n_iter_max=100):
+    def __init__(self, metric='riemann', threshold=3, n_iter_max=100,
+                 pos_label=1, neg_label=0):
         """Init."""
         self.metric = metric
         self.threshold = threshold
         self.n_iter_max = n_iter_max
+        if pos_label == neg_label:
+            raise(ValueError("Positive and Negative labels must be different"))
+        self.pos_label = pos_label
+        self.neg_label = neg_label
 
     def fit(self, X, y=None):
         """Fit the potato from covariance matrices.
@@ -308,15 +319,15 @@ class Potato(BaseEstimator, TransformerMixin, ClassifierMixin):
             if len(y) != len(X):
                 raise ValueError('y must be the same lenght of X')
 
-            classes = numpy.unique(y)
+            classes = numpy.int32(numpy.unique(y))
 
             if len(classes) > 2:
                 raise ValueError('number of classes must be maximum 2')
 
-            if 1 not in classes:
+            if self.pos_label not in classes:
                 raise ValueError('y must contain a positive class')
 
-            y_old = numpy.array(y)
+            y_old = numpy.int32(numpy.array(y) == self.pos_label)
         else:
             y_old = numpy.ones(len(X))
         # start loop
@@ -368,7 +379,9 @@ class Potato(BaseEstimator, TransformerMixin, ClassifierMixin):
         """
         z = self.transform(X)
         pred = z < self.threshold
-        return pred
+        out = numpy.zeros_like(z) + self.neg_label
+        out[pred] = self.pos_label
+        return out
 
     def _get_z_score(self, d):
         """get z score from distance."""