+++ tests/test_utils_base.py
@@ -1,6 +1,7 @@
-from numpy.testing import assert_array_almost_equal
 import numpy as np
-
+from nose.tools import assert_raises
+from numpy.testing import assert_array_almost_equal
+from pyriemann.utils.mean import mean_riemann
 from pyriemann.utils.base import (sqrtm, invsqrtm, logm, expm, powm)
 
 
@@ -37,3 +38,11 @@ def test_powm():
     C = 2*np.eye(3)
     Ctrue = (2**0.5)*np.eye(3)
     assert_array_almost_equal(powm(C, 0.5), Ctrue)
+
+
+def test_check_raise():
+    """Test chech SPD matrices"""
+    C = 2*np.ones((10, 3, 3))
+    # This is an indirect check, the riemannian mean must crash when the
+    # matrices are not SPD.
+    assert_raises(ValueError, mean_riemann, C)