+++ doc/whatsnew.rst
@@ -20,6 +20,10 @@ v0.2.5 (Future release)
 
 - Support for python 3.4 in travis
 
+- Added Shrinkage transformer
+
+- Added Coherences transformer
+
 v0.2.4 (June 2016)
 -------------------